+++ backend/substrapp/tasks/__init__.py
@@ -1,6 +1,5 @@
-from substrapp.tasks.tasks_compute_plan import delete_cp_pod_and_dirs_and_optionally_images
+from substrapp.tasks.tasks_compute_plan import delete_cp_pod_and_dirs
 from substrapp.tasks.tasks_compute_task import compute_task
-from substrapp.tasks.tasks_docker_registry import clean_old_images_task
 from substrapp.tasks.tasks_docker_registry import docker_registry_garbage_collector_task
 from substrapp.tasks.tasks_outputs import remove_transient_outputs_from_orc
 from substrapp.tasks.tasks_remove_intermediary_models import remove_intermediary_model_from_db
@@ -8,9 +7,8 @@ from substrapp.tasks.tasks_remove_intermediary_models import remove_intermediary
 from substrapp.tasks.tasks_save_image import save_image_task
 
 __all__ = [
-    "delete_cp_pod_and_dirs_and_optionally_images",
+    "delete_cp_pod_and_dirs",
     "compute_task",
-    "clean_old_images_task",
     "docker_registry_garbage_collector_task",
     "remove_intermediary_models_from_buffer",
     "remove_transient_outputs_from_orc",